オフィスマスターのリファクタリング

+++ phalanx/app/Http/Controllers/OfficeController.php
@@ -13,6 +13,7 @@ use App\Models\Office;
 use App\Models\ChatRoom;
 use Illuminate\Validation\Rule;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use App\Http\Requests\OfficeRequest;
 use Illuminate\Support\Facades\Validator;
 use Carbon\Carbon;
@@ -44,51 +45,48 @@ class OfficeController extends Controller
     }
 
      /**
-     * 新規データ作成の実行部分
+     * 新規データ作成の受け取り部分
      */
-    public function store(Request $request)
+    public function store(OfficeRequest $request)
     {
-        $request->validate([
-            'office_name' => 'required|unique:offices,office_name',
-            'sort' => 'required|unique:offices,sort'
-        ],
-        [
-            'office_name.required' => '事業所名を入力してください。',
-            'sort.required' => '表示する順番を数字で入力してください。',
-            'office_name.unique' => 'その事業所は既に登録されています。',
-            'sort.unique' => 'その番号は既に登録されています。'
-        ]);
-
-
-        //ログイン中のユーザーデータを取得
-        $user = Auth::user();
-
-        //各種リクエストのデータを取得
-        $officename = $request->input("office_name");
-        $sort = $request->input("sort");
-
-        //現在時刻の取得
-        $now = Carbon::now()->isoFormat('YYYY-MM-DD HH:mm:ss');
-
-        //Officeインスタンスを作成、各種データを挿入後登録
-        $office = new Office();
-        $office->office_name = $officename;
-        $office->sort = $sort;
-        $office->create_user_id = isset($user->id);
-        $office->update_user_id = isset($user->id);
-        $office->created_at = $now;
-        $office->updated_at = $now;
-        $office->save();
-
-        $con = app()->make("App\Http\Controllers\ChatRoomController");
-        $id = $con->storeDetail(new Request([
-            "room_title" => $office->office_name . "職員",
-            "distinction_number" => 1,
-            "office_id" => $office->id,
-        ]));
+        $id = $this->storeDetail($request);
 
         return redirect()->route("office.index");
     }
+
+    /**
+     * 新規データ作成の実行部分
+     * 
+     * @param Request $request
+     * @return int $id
+     */
+    public function storeDetail(Request $request) {
+
+        $office = null;
+        DB::transaction(function () use(&$office, $request) {
+            $now = Carbon::now()->isoFormat('YYYY-MM-DD HH:mm:ss');
+            $office = Office::create([
+                "office_name" => $request->input("office_name"),
+                "en_office_name" => $request->input("en_office_name"),
+                "sort" => $request->input("sort"),
+                "create_user_id" => Auth::id(),
+                "update_user_id" => Auth::id(),
+                "created_at" => $now,
+                "update_at" => $now,
+            ]);
+
+            $con = app()->make("App\Http\Controllers\ChatRoomController");
+            $id = $con->storeDetail(new Request([
+                "room_title" => $office->office_name . "職員",
+                "distinction_number" => 1,
+                "office_id" => $office->id,
+            ]));
+        });
+
+        if(isset($office)) return $office->id;
+        return -1;
+    }
+
 
     /**
      * データの編集
@@ -109,43 +107,47 @@ class OfficeController extends Controller
      */
 
     /**
-     * データ編集の実行部分
+     * データ編集の受け取り部分
      */
-    public function update(Request $request, $id)
+    public function update(OfficeRequest $request, $id)
     {
-        $request->validate([
-            'office_name' => ['required',Rule::unique('offices')->ignore($id)],
-            'sort' => ['required',Rule::unique('offices')->ignore($id)]
-        ],
-        [
-            'office_name.required' => '事業所名を入力してください。',
-            'sort.required' => '表示する順番を数字で入力してください。',
-            'office_name.unique' => 'その事業所は既に登録されています。',
-            'sort.unique' => 'その番号は既に登録されています。'
-        ]);
-
-        //ログイン中のユーザーデータを取得
-        $user = Auth::user();
-
-        //各種リクエストのデータを取得
-        $officename = $request->input("office_name");
-        $sort = $request->input("sort");
-
-    //現在時刻を取得
-        $now = Carbon::now()->isoFormat('YYYY-MM-DD HH:mm:ss');
-
-        $office = Office::findOrFail($id);
-        $office->office_name = $officename;
-        $office->sort = $sort;
-        $office->update_user_id = $user->id;
-        $office->updated_at = $now;
-        $office->save();
-
-        ChatRoom::whereNull("deleted_at")->where("distinction_number", 1)->where("office_id", $office->id)->update([
-            "room_title" => $office->office_name . "職員"
-        ]);
+        $id = $this->updateDetail($request, $id);
 
         return redirect()->route("office.index");
+    }
+
+    /**
+     * データ更新の実行部分
+     * 
+     * @param Request $request, int $id
+     * @return int $id
+     */
+    public function updateDetail(Request $request, $id) {
+
+        $office = null;
+        DB::transaction(function () use(&$office, $request, $id) {
+            $now = Carbon::now()->isoFormat('YYYY-MM-DD HH:mm:ss');
+            $office = Office::whereNull("deleted_at")->where("id", $id)->first();
+
+            if(isset($office)) {
+                $office->update([
+                    "office_name" => $request->input("office_name"),
+                    "en_office_name" => $request->input("en_office_name"),
+                    "sort" => $request->input("sort"),
+                    "update_user_id" => Auth::id(),
+                    "updated_at" => $now
+                ]);
+
+                ChatRoom::whereNull("deleted_at")->where("distinction_number", 1)->where("office_id", $office->id)->update([
+                    "room_title" => $office->office_name . "職員",
+                    "update_user_id" => Auth::id(),
+                    "updated_at" => $now
+                ]);
+            }
+        });
+
+        if(isset($office)) return $id;
+        return -1;
     }
 
     /**
@@ -153,20 +155,22 @@ class OfficeController extends Controller
      */
     public function destroy($id)
     {
-        //ログイン中のユーザーデータを取得
-        $user = Auth::user();
+        DB::transaction(function () use($id) {
+            $now = Carbon::now()->isoFormat('YYYY-MM-DD HH:mm:ss');
+            $office = whereNull("deleted_at")->where("id", $id);
 
-        //現在時刻を取得
-        $now = Carbon::now()->isoFormat('YYYY-MM-DD HH:mm:ss');
+            if(isset($office)) {
+                $office->update([
+                    "delete_user_id" => Auth::id(),
+                    "deleted_at" => $now
+                ]);
 
-        //削除を実行
-        $office = Office::findOrFail($id);
+                $con = app()->make("App\Http\Controllers\ChatRoomController");
+                $chatRoom = ChatRoom::whereNull("deleted_at")->where("distinction_number", 1)->where("office_id", $office->id)->first();
+                $con->destroy($chatRoom->id);
+            }
+        });
 
-        $con = app()->make("App\Http\Controllers\ChatRoomController");
-        $chatRoom = ChatRoom::whereNull("deleted_at")->where("distinction_number", 1)->where("office_id", $office->id)->first();
-        $con->destroy($chatRoom->id);
-
-        $office->delete();
         return redirect()->route("office.index");
     }
